+++ packages/slate-hyperscript/src/index.js
@@ -8,7 +8,7 @@ import {
   Inline,
   Mark,
   Node,
-  Selection,
+  Range,
   State,
   Text
 } from 'slate'
@@ -71,13 +71,13 @@ const CREATORS = {
   },
 
   selection(tagName, attributes, children) {
-    return Selection.create(attributes)
+    return Range.create(attributes)
   },
 
   state(tagName, attributes, children) {
     const { data } = attributes
     const document = children.find(Document.isDocument)
-    let selection = children.find(Selection.isSelection) || Selection.create()
+    let selection = children.find(Range.isRange) || Range.create()
     const props = {}
 
     // Search the document's texts to see if any of them have the anchor or
