+++ packages/slate/test/schemas/core/document-no-inline-children.js
@@ -31,9 +31,9 @@ export const output = {
         nodes: [
           {
             kind: 'text',
-            ranges: [
+            leaves: [
               {
-                kind: 'range',
+                kind: 'leaf',
                 text: 'two',
                 marks: [],
               }
