+++ packages/slate/test/schemas/core/document-no-text-children.js
@@ -29,9 +29,9 @@ export const output = {
         nodes: [
           {
             kind: 'text',
-            ranges: [
+            leaves: [
               {
-                kind: 'range',
+                kind: 'leaf',
                 text: 'two',
                 marks: [],
               }
