+++ packages/slate/test/schemas/core/remove-empty-inline.js
@@ -28,9 +28,9 @@ export const output = {
         nodes: [
           {
             kind: 'text',
-            ranges: [
+            leaves: [
               {
-                kind: 'range',
+                kind: 'leaf',
                 text: '',
                 marks: [],
               }
