+++ packages/slate/test/serializers/raw/deserialize/block-nested.js
@@ -22,10 +22,10 @@ export const input = {
             nodes: [
               {
                 kind: 'text',
-                ranges: [
+                leaves: [
                   {
                     text: 'one',
-                    kind: 'range',
+                    kind: 'leaf',
                     marks: []
                   }
                 ]
