+++ packages/slate/test/serializers/raw/deserialize/block-with-data.js
@@ -18,9 +18,9 @@ export const input = {
         nodes: [
           {
             kind: 'text',
-            ranges: [
+            leaves: [
               {
-                kind: 'range',
+                kind: 'leaf',
                 text: 'one',
                 marks: []
               }
