+++ packages/slate/test/serializers/raw/serialize/block-nested.js
@@ -34,10 +34,10 @@ export const output = {
             nodes: [
               {
                 kind: 'text',
-                ranges: [
+                leaves: [
                   {
                     text: 'one',
-                    kind: 'range',
+                    kind: 'leaf',
                     marks: []
                   }
                 ]
