+++ packages/slate/test/serializers/raw/serialize/block-with-data.js
@@ -28,9 +28,9 @@ export const output = {
         nodes: [
           {
             kind: 'text',
-            ranges: [
+            leaves: [
               {
-                kind: 'range',
+                kind: 'leaf',
                 text: 'one',
                 marks: []
               }
