+++ packages/slate/test/serializers/raw/serialize/block-with-is-void.js
@@ -24,9 +24,9 @@ export const output = {
         nodes: [
           {
             kind: 'text',
-            ranges: [
+            leaves: [
               {
-                kind: 'range',
+                kind: 'leaf',
                 text: ' ',
                 marks: [],
               }
