+++ packages/slate/test/serializers/raw/serialize/block.js
@@ -26,9 +26,9 @@ export const output = {
         nodes: [
           {
             kind: 'text',
-            ranges: [
+            leaves: [
               {
-                kind: 'range',
+                kind: 'leaf',
                 text: 'one',
                 marks: [],
               }
