+++ packages/slate/test/serializers/raw/serialize/preserve-data.js
@@ -27,10 +27,10 @@ export const output = {
         nodes: [
           {
             kind: 'text',
-            ranges: [
+            leaves: [
               {
                 text: 'one',
-                kind: 'range',
+                kind: 'leaf',
                 marks: []
               }
             ]
