+++ packages/slate/test/serializers/raw/serialize/preserve-selection.js
@@ -26,10 +26,10 @@ export const output = {
         nodes: [
           {
             kind: 'text',
-            ranges: [
+            leaves: [
               {
                 text: 'one',
-                kind: 'range',
+                kind: 'leaf',
                 marks: []
               }
             ]
@@ -39,7 +39,7 @@ export const output = {
     ]
   },
   selection: {
-    kind: 'selection',
+    kind: 'range',
     anchorPath: [0, 0],
     anchorOffset: 0,
     focusPath: [0, 0],
