+++ packages/slate/test/serializers/raw/serialize/range-with-mark.js
@@ -26,14 +26,14 @@ export const output = {
         nodes: [
           {
             kind: 'text',
-            ranges: [
+            leaves: [
               {
-                kind: 'range',
+                kind: 'leaf',
                 text: 'o',
                 marks: [],
               },
               {
-                kind: 'range',
+                kind: 'leaf',
                 text: 'n',
                 marks: [
                   {
@@ -44,7 +44,7 @@ export const output = {
                 ]
               },
               {
-                kind: 'range',
+                kind: 'leaf',
                 text: 'e',
                 marks: [],
               }
